move roi logic to explorer
